Feat: Add search parameter to url Feat: Synonym and antonym can be clicked to search for the word Refactor: Rename variables to be more descriptive
Users can now utilize the search bar to search for a word and the word will be added to the url. The url will be used to fetch the definition of the word. The word will also be added to the url when the user clicks on a synonym or antonym in the list of words.